requirements.txt
Django==1.8
django-authtools==1.2.0
django-braces==1.8.0
django-extensions==1.5.2
django-pipeline==1.4.7
djangorestframework
futures==2.2.0
six==1.9.0
